AVRO-4223
Move reusable methods to abstract class
